Refactor category views
- audit moved to ideax.util
