+++ ideax/ideax/tests/views/test_category.py
@@ -22,8 +22,8 @@ class TestCategoryNew:
             category_new(request)
 
     def test_get(self, rf, factory_user, mocker):
-        render = mocker.patch('ideax.ideax.views.render')
-        form = mocker.patch('ideax.ideax.views.CategoryForm')
+        render = mocker.patch('ideax.ideax.views.category.render')
+        form = mocker.patch('ideax.ideax.views.category.CategoryForm')
 
         request = rf.get('/')
         request.user = factory_user('ideax.add_category')
@@ -33,8 +33,8 @@ class TestCategoryNew:
         render.assert_called_once_with(request, 'ideax/category_new.html', {'form': form.return_value})
 
     def test_post(self, rf, factory_user, mocker, messages):
-        audit = mocker.patch('ideax.ideax.views.audit')
-        form = mocker.patch('ideax.ideax.views.CategoryForm')
+        audit = mocker.patch('ideax.ideax.views.category.audit')
+        form = mocker.patch('ideax.ideax.views.category.CategoryForm')
         form.return_value.is_valid.return_value = True
         user_profile = mocker.patch('ideax.users.models.UserProfile.objects')
         user_profile.get.return_value = None
@@ -51,7 +51,7 @@ class TestCategoryNew:
         assert messages.messages == ['Category saved successfully!']
 
     def test_post_invalid_form(self, rf, factory_user, mocker):
-        form = mocker.patch('ideax.ideax.views.CategoryForm')
+        form = mocker.patch('ideax.ideax.views.category.CategoryForm')
         form.return_value.is_valid.return_value = False
         user_profile = mocker.patch('ideax.users.models.UserProfile.objects')
         user_profile.get.return_value = None
@@ -84,9 +84,9 @@ class TestCategoryEdit:
             category_edit(request, 1)
 
     def test_get(self, rf, factory_user, mocker):
-        get = mocker.patch('ideax.ideax.views.get_object_or_404')
-        render = mocker.patch('ideax.ideax.views.render')
-        form = mocker.patch('ideax.ideax.views.CategoryForm')
+        get = mocker.patch('ideax.ideax.views.category.get_object_or_404')
+        render = mocker.patch('ideax.ideax.views.category.render')
+        form = mocker.patch('ideax.ideax.views.category.CategoryForm')
 
         request = rf.get(f'/category/55/edit/')
         request.user = factory_user('ideax.change_category')
@@ -96,9 +96,9 @@ class TestCategoryEdit:
         render.assert_called_once_with(request, 'ideax/category_edit.html', {'form': form.return_value})
 
     def test_post(self, rf, factory_user, mocker, messages):
-        get = mocker.patch('ideax.ideax.views.get_object_or_404')
-        mocker.patch('ideax.ideax.views.audit')
-        category_form = mocker.patch('ideax.ideax.views.CategoryForm')
+        get = mocker.patch('ideax.ideax.views.category.get_object_or_404')
+        mocker.patch('ideax.ideax.views.category.audit')
+        category_form = mocker.patch('ideax.ideax.views.category.CategoryForm')
         category_form.return_value.is_valid.return_value = True
 
         request = rf.post(f'/category/55/edit/', {})
@@ -113,11 +113,11 @@ class TestCategoryEdit:
         assert messages.messages == ['Category changed successfully!']
 
     def test_post_invalid_form(self, rf, factory_user, mocker, messages):
-        get = mocker.patch('ideax.ideax.views.get_object_or_404')
-        mocker.patch('ideax.ideax.views.audit')
-        category_form = mocker.patch('ideax.ideax.views.CategoryForm')
+        get = mocker.patch('ideax.ideax.views.category.get_object_or_404')
+        mocker.patch('ideax.ideax.views.category.audit')
+        category_form = mocker.patch('ideax.ideax.views.category.CategoryForm')
         category_form.return_value.is_valid.return_value = False
-        render = mocker.patch('ideax.ideax.views.render')
+        render = mocker.patch('ideax.ideax.views.category.render')
 
         request = rf.post(f'/category/55/edit/', {})
         request.user = factory_user('ideax.change_category')
@@ -144,11 +144,11 @@ class TestCategoryRemove:
             category_remove(request, 1)
 
     def test_get(self, rf, factory_user, mocker, messages):
-        audit = mocker.patch('ideax.ideax.views.audit')
-        get = mocker.patch('ideax.ideax.views.get_object_or_404')
-        get_category_list = mocker.patch('ideax.ideax.views.get_category_list')
+        audit = mocker.patch('ideax.ideax.views.category.audit')
+        get = mocker.patch('ideax.ideax.views.category.get_object_or_404')
+        get_category_list = mocker.patch('ideax.ideax.views.category.get_category_list')
         get_category_list.return_value = {}
-        category = mocker.patch('ideax.ideax.views.Category')
+        category = mocker.patch('ideax.ideax.views.category.Category')
         category.__name__ = 'Category'
 
         request = rf.get('/')
@@ -173,9 +173,9 @@ class TestCategoryList:
 
     def test_get(self, rf, mocker, common_user):
         mocker.patch('ideax.ideax.views.audit')
-        get_category_list = mocker.patch('ideax.ideax.views.get_category_list')
+        get_category_list = mocker.patch('ideax.ideax.views.category.get_category_list')
         get_category_list.return_value = {}
-        render = mocker.patch('ideax.ideax.views.render')
+        render = mocker.patch('ideax.ideax.views.category.render')
 
         request = rf.get('/')
         request.user = common_user
