+++ ideax/ideax/views/__init__.py
@@ -3,7 +3,6 @@ import json
 import uuid
 import collections
 import mistune
-import logging
 import csv
 
 from datetime import date
@@ -31,31 +30,16 @@ from ..models import (
     Category, Comment, Dimension, Evaluation, Category_Image, Use_Term, Challenge
 )
 from ..forms import (
-    IdeaForm, CriterionForm, CategoryForm, CategoryImageForm,
+    IdeaForm, CriterionForm, CategoryImageForm,
     EvaluationForm, ChallengeForm, UseTermForm, DimensionForm
 )
 from ...singleton import Profanity_Check
 from ...mail_util import MailUtil
-from ...util import get_ip, get_client_ip
+from ...util import get_ip, get_client_ip, audit
 
-
-# Creating log object
-logger = logging.getLogger('audit_log')
+from .category import category_edit, category_list, category_new, category_remove, get_category_list
 
 mail_util = MailUtil()
-
-
-def audit(username, ip_addr, operation, className, objectId):
-    logger.info(
-        '%(username)s|%(ip_addr)s|%(operation)s|%(className)s|%(objectId)s',
-        {
-            'username': username,
-            'ip_addr': ip_addr,
-            'operation': operation,
-            'className': className,
-            'objectId': objectId
-        }
-    )
 
 
 def index(request):
@@ -421,68 +405,6 @@ def criterion_remove(request, pk):
     audit(request.user.username, get_client_ip(request), 'REMOVE_CRITERION_SAVE', Criterion.__name__, str(pk))
 
     return redirect('criterion_list')
-
-
-@login_required
-@permission_required('ideax.add_category', raise_exception=True)
-def category_new(request):
-    if request.method == "POST":
-        form = CategoryForm(request.POST)
-        if form.is_valid():
-            category = form.save(commit=False)
-            category.author = UserProfile.objects.get(user=request.user)
-            category.creation_date = timezone.now()
-            category.save()
-            messages.success(request, _('Category saved successfully!'))
-            audit(request.user.username, get_client_ip(request), 'CREATE_CATEGORY', Category.__name__, category.id)
-            return redirect('category_list')
-    else:
-        form = CategoryForm()
-    return render(request, 'ideax/category_new.html', {'form': form})
-
-
-@login_required
-@permission_required('ideax.change_category', raise_exception=True)
-def category_edit(request, pk):
-    category = get_object_or_404(Category, pk=pk)
-    if request.method == "POST":
-        form = CategoryForm(request.POST, instance=category)
-        if form.is_valid():
-            form.save()
-            messages.success(request, _('Category changed successfully!'))
-            audit(
-                request.user.username,
-                get_client_ip(request),
-                'EDIT_CATEGORY_SAVE',
-                Category.__name__,
-                str(category.id)
-            )
-            return redirect('category_list')
-    else:
-        form = CategoryForm(instance=category)
-
-    return render(request, 'ideax/category_edit.html', {'form': form})
-
-
-@login_required
-@permission_required('ideax.delete_category', raise_exception=True)
-def category_remove(request, pk):
-    category = get_object_or_404(Category, pk=pk)
-    category.discarded = True
-    category.save()
-    messages.success(request, _('Category removed successfully!'))
-    audit(request.user.username, get_client_ip(request), 'REMOVE_CATEGORY', Category.__name__, str(pk))
-    return redirect('category_list')
-
-
-@login_required
-def category_list(request):
-    audit(request.user.username, get_client_ip(request), 'CATEGORY_LIST', Category.__name__, '')
-    return render(request, 'ideax/category_list.html', get_category_list())
-
-
-def get_category_list():
-    return {'category_list': Category.objects.filter(discarded=False)}
 
 
 @login_required
@@ -1087,3 +1009,8 @@ def dimension_remove(request, pk):
     messages.success(request, _('Dimension removed successfully!'))
     audit(request.user.username, get_client_ip(request), 'REMOVE_DIMENSION', Dimension.__name__, str(pk))
     return redirect('dimension_list')
+
+
+__all__ = [
+    'category_edit', 'category_list', 'category_new', 'category_remove', 'get_category_list'
+]
